imagemngr.py: repull enqueued images (#3)
This is the change that we implemented on Blue Waters to fix the problem where some images get stuck in the ENQUEUED state and can not be pulled again. The most likely reason is that when Shifter checks whether an ENQUEUED image is pullable it concludes that it's not because no condition in the pullable function is met to return True. This is likely to happen because ENQUEUED pull records don't have all the entries (such as last_heartbeat).